+++ src/Layer.cpp
@@ -10,10 +10,10 @@ constexpr auto f_bias{1};
 constexpr auto f_weightInitMinValue{-1.0};
 constexpr auto f_weightInitMaxValue{+1.0};
 
-inline auto randomValue(double min, double max) {
+inline auto randomValue(float min, float max) {
   std::random_device rd;
   std::mt19937 eng{rd()};
-  std::uniform_real_distribution<double> dist{min, max};
+  std::uniform_real_distribution<float> dist{min, max};
   return dist(eng);
 }
 } // namespace
@@ -24,11 +24,11 @@ Layer::Layer(std::size_t id, std::size_t numberOfInputs,
     : m_id{id}, m_numberOfInputs{numberOfInputs},
       m_numberOfNeurons{numberOfNeurons},
       m_activationFunction{ActivationFunction::instance(activationFunction)},
-      m_intermediateQtys{Eigen::VectorXd::Zero(numberOfNeurons)},
-      m_inputs{Eigen::VectorXd::Ones(numberOfInputs + f_bias)},
-      m_outputs{Eigen::VectorXd::Zero(numberOfNeurons)},
-      m_errors{Eigen::VectorXd::Zero(numberOfNeurons)},
-      m_weights{Eigen::MatrixXd::NullaryExpr(
+      m_intermediateQtys{Eigen::VectorXf::Zero(numberOfNeurons)},
+      m_inputs{Eigen::VectorXf::Ones(numberOfInputs + f_bias)},
+      m_outputs{Eigen::VectorXf::Zero(numberOfNeurons)},
+      m_errors{Eigen::VectorXf::Zero(numberOfNeurons)},
+      m_weights{Eigen::MatrixXf::NullaryExpr(
           numberOfInputs + f_bias, numberOfNeurons,
           []() {
             return randomValue(f_weightInitMinValue, f_weightInitMaxValue);
@@ -41,19 +41,19 @@ std::size_t Layer::numberOfInputs() const { return m_numberOfInputs; }
 
 std::size_t Layer::numberOfNeurons() const { return m_numberOfNeurons; }
 
-double Layer::loss() const { return m_loss; }
+float Layer::loss() const { return m_loss; }
 
-const Eigen::MatrixXd &Layer::weights() const { return m_weights; }
+const Eigen::MatrixXf &Layer::weights() const { return m_weights; }
 
-const Eigen::VectorXd &Layer::outputs() const { return m_outputs; }
+const Eigen::VectorXf &Layer::outputs() const { return m_outputs; }
 
-const Eigen::VectorXd &Layer::errors() const { return m_errors; }
+const Eigen::VectorXf &Layer::errors() const { return m_errors; }
 
-Eigen::MatrixXd Layer::computeGradients() const {
+Eigen::MatrixXf Layer::computeGradients() const {
   return m_inputs * m_errors.transpose();
 }
 
-void Layer::updateOutputs(const Eigen::VectorXd &inputs) {
+void Layer::updateOutputs(const Eigen::VectorXf &inputs) {
   m_inputs.head(inputs.size()) = inputs;
   m_intermediateQtys = m_weights.transpose() * m_inputs;
   std::transform(
@@ -63,7 +63,8 @@ void Layer::updateOutputs(const Eigen::VectorXd &inputs) {
 
 void Layer::updateErrors(const Layer &nextLayer) {
   auto &nextLayerWeights{nextLayer.weights()};
-  m_errors = nextLayerWeights.topRows(nextLayerWeights.rows() - 1) * nextLayer.errors();
+  m_errors = nextLayerWeights.topRows(nextLayerWeights.rows() - 1) *
+             nextLayer.errors();
   std::transform(m_errors.cbegin(), m_errors.cend(),
                  m_intermediateQtys.cbegin(), m_errors.begin(),
                  [this](auto e, auto q) {
@@ -71,7 +72,7 @@ void Layer::updateErrors(const Layer &nextLayer) {
                  });
 }
 
-void Layer::updateErrors(const Eigen::VectorXd &targets,
+void Layer::updateErrors(const Eigen::VectorXf &targets,
                          const CostFunction &costFunction) {
   std::transform(m_outputs.cbegin(), m_outputs.cend(), targets.cbegin(),
                  m_errors.begin(), [&costFunction](auto o, auto t) {
@@ -82,10 +83,10 @@ void Layer::updateErrors(const Eigen::VectorXd &targets,
   m_loss = losses.sum() / m_numberOfNeurons;
 }
 
-void Layer::updateWeights(double learnRate) {
+void Layer::updateWeights(float learnRate) {
   updateWeights(computeGradients(), learnRate);
 }
 
-void Layer::updateWeights(const Eigen::MatrixXd &gradients, double learnRate) {
+void Layer::updateWeights(const Eigen::MatrixXf &gradients, float learnRate) {
   m_weights -= learnRate * gradients;
 }
